fix(SpotlightPopoverTour, PopupArrow): revert the arrow tip radius and even out the arrow-to-spotlight gap
Raising `tipRadius` to `xsmall` read as too rounded in practice — the ratio scales
against the arrow's own dimensions, so it bowed the edges rather than just blunting
the tip. Restored to `2xsmall`, which also takes the `Popover` and `Tooltip`
snapshots back to their pre-existing output.

The gap between the arrow tip and the spotlight also varied per step, because the
mask resolved the step's ref to the painted component while `TourPopover` anchored
to the raw ref. Any wrapper overhang landed in that gap. Both now resolve through
`resolveSpotlightTarget`, moved into the shared web utils.
