+++ packages/blade/src/components/PopupArrow/PopupArrow.web.tsx
@@ -17,12 +17,11 @@ const PopupArrow = React.forwardRef<SVGSVGElement, PopupArrowProps>(
         strokeWidth={strokeWidth ?? 0}
         /**
          * `tipRadius` is a ratio, not a px value — FloatingArrow scales it against the arrow's
-         * own width/height. At `2xsmall` (2) a 22x12 arrow is blunted by only 1.5px, which
-         * disappears into antialiasing at 1x and reads as a hard point next to the popup's
-         * rounded corners. `xsmall` (4) is the largest value that still leaves straight edges —
-         * the shape degenerates into a lens by 8.
+         * own width/height, so the rounding grows much faster than the token number suggests.
+         * `xsmall` (4) visibly bows the arrow's edges; `2xsmall` (2) softens the tip while
+         * keeping the sides straight.
          */
-        tipRadius={theme.border.radius.xsmall}
+        tipRadius={theme.border.radius['2xsmall']}
         style={{ ...style }}
       />
     );
