+++ packages/blade/src/components/SpotlightPopoverTour/Tour.web.tsx
@@ -7,6 +7,8 @@ import { TourContext } from './TourContext';
 import type { TourElement } from './TourContext';
 import { TourPopover } from './TourPopover';
 import {
+  readBorderRadius,
+  resolveSpotlightTarget,
   smoothScroll,
   useDelayedState,
   useIntersectionObserver,
@@ -21,66 +23,6 @@ import { isBrowser } from '~utils';
 import { useIsomorphicLayoutEffect } from '~utils/useIsomorphicLayoutEffect';
 
 const asHTMLElement = (el: TourElement | null): HTMLElement | null => el as HTMLElement | null;
-
-const readBorderRadius = (el: HTMLElement): number => {
-  const { borderTopLeftRadius } = window.getComputedStyle(el);
-  const parsed = Number.parseFloat(borderTopLeftRadius);
-
-  if (Number.isNaN(parsed)) return 0;
-  // Percentage radii (e.g. `50%` on a circular avatar) resolve against the box itself.
-  if (borderTopLeftRadius.includes('%')) {
-    const { width, height } = el.getBoundingClientRect();
-    return (Math.min(width, height) * parsed) / 100;
-  }
-
-  return parsed;
-};
-
-// `getComputedStyle` reports unset properties as an empty string in some environments
-// (notably jsdom), so treat "unset" the same as "not painted" rather than as a value.
-const isPainted = (value: string): boolean =>
-  Boolean(value) && value !== 'none' && value !== 'transparent' && value !== 'rgba(0, 0, 0, 0)';
-
-/** Whether an element paints anything of its own, or is a bare layout shell. */
-const drawsOwnShape = (el: HTMLElement): boolean => {
-  const { borderTopWidth, backgroundColor, backgroundImage } = window.getComputedStyle(el);
-  return (
-    readBorderRadius(el) > 0 ||
-    Number.parseFloat(borderTopWidth) > 0 ||
-    isPainted(backgroundColor) ||
-    isPainted(backgroundImage)
-  );
-};
-
-/**
- * Resolves the element the spotlight should actually trace.
- *
- * `SpotlightPopoverTourStep` clones its child to attach a ref, so consumers commonly wrap
- * their UI in a layout element just to forward one. Measuring that wrapper is wrong twice
- * over: it has no corner radius to inherit, and a wrapper stretched by its parent's layout
- * (e.g. `alignItems="stretch"` in a flex row) is taller than the component inside it, which
- * makes the spotlight's padding look uneven — 6px on three sides and more at the bottom.
- *
- * So when the step's element paints nothing itself and simply wraps a single child that
- * fills it on at least one axis, we trace that child instead. The fill check keeps us from
- * shrinking onto an inner element that merely happens to be first — a wrapper around
- * narrower content is still measured as the wrapper.
- */
-const resolveSpotlightTarget = (element: HTMLElement): HTMLElement => {
-  if (drawsOwnShape(element)) return element;
-
-  const { children } = element;
-  if (children.length !== 1) return element;
-
-  const child = children[0] as HTMLElement;
-  const parentRect = element.getBoundingClientRect();
-  const childRect = child.getBoundingClientRect();
-  const fillsAnAxis =
-    Math.abs(childRect.width - parentRect.width) <= 1 ||
-    Math.abs(childRect.height - parentRect.height) <= 1;
-
-  return fillsAnAxis ? child : element;
-};
 
 const SpotlightPopoverTour = ({
   steps,
